Pie chart now working with labels and tooltips. Uncomment in app.js line 119 to render. Still needs css styling.
